requirements.txt
lxml==5.3.2
python-docx==1.1.2
setuptools==79.0.0
typing_extensions==4.13.2
